+++ sunnypilot/models/helpers.py
